绘制棋盘
